Added Summer Time Warning bit
The Summer Time bit is transmitted during the first minute before Summer Time actually begins. This works out to be only 2 seconds before Summer Time begins since this is the 59th bit of 60. This allows clocks to make the change on time and not have to wait another minute for the time to be broadcast again.
The Summer Time Warning bit is only transmitted during the hour before the Summer Time bit is set, working out to be 1 hour and 7 seconds before the change to Summer Time is actually made.